api/requirements.txt
fastapi>=0.100.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
httpx>=0.24.0
uvicorn>=0.24.0
